fix: PHP 8.5 compatibility - const qualifier and return type cast
- runkit_functions.c: Add const qualifier to atomic_type pointer
  to fix 'discarded-qualifiers' error with PHP 8.5 ZEND_TYPE_LIST_FOREACH macro

- runkit_constants.c: Cast zend_register_constant() return to int
  to fix 'makes integer from pointer' error (PHP 8.5 changed return type
  from int to zend_constant*)

+++ runkit_constants.c
@@ -408,7 +408,7 @@ static int php_runkit_global_constant_add(zend_string *constname, zval *value)
 
 	c.name = constname;
 	ZEND_CONSTANT_SET_FLAGS(&c, CONST_CS, PHP_USER_CONSTANT);
-	return zend_register_constant(&c);
+	return (int)zend_register_constant(&c);
 }
 /* }}} */
